Commonized the various configs
