+++ src/config/DataConfig.ts
@@ -1,0 +1,11 @@
+import { CharacterData } from "../data/CharacterData"
+
+declare global
+{
+    interface DataConfig
+    {
+        Actor: { type: 'character', data: CharacterData }
+    }
+}
+
+export { }
