+++ src/config/DocumentClassConfig.ts
@@ -1,0 +1,11 @@
+import StatusEffect from '../effects/StatusEffect'
+
+declare global
+{
+    interface DocumentClassConfig
+    {
+        ActiveEffect: typeof StatusEffect
+    }
+}
+
+export { }
