--- src/data/config.ts
@@ -1,13 +1,0 @@
-import { CharacterData } from "./CharacterData"
-
-declare global
-{
-    interface SourceConfig
-    {
-        Actor: { type: 'character', data: CharacterData }
-    }
-
-    interface DataConfig extends SourceConfig
-    {
-    }
-}
